move WorkspaceVariableDisplayValue to ark
